wedge_utils_{avx2,sse2}: fix implicit conversion warnings
w/clang -fsanitize=integer quiets warnings of the form:
implicit conversion from type 'uint64_t' (aka 'unsigned long') of value
18446744073666365769 (64-bit, unsigned) to type 'int64_t' (aka 'long')
changed the value to -43185847 (64-bit, signed)

Bug: aomedia:3136
Bug: b/229626362

+++ av1/encoder/x86/wedge_utils_avx2.c
@@ -142,7 +142,7 @@ int8_t av1_wedge_sign_from_residuals_avx2(const int16_t *ds, const uint8_t *m,
   v_acc_q_0 = _mm_add_epi64(v_acc_q_0, v_acc_q_1);
 
 #if ARCH_X86_64
-  acc = (uint64_t)_mm_extract_epi64(v_acc_q_0, 0);
+  acc = _mm_extract_epi64(v_acc_q_0, 0);
 #else
   xx_storel_64(&acc, v_acc_q_0);
 #endif
